docker/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
huggingface-hub==0.20.3
llama-cpp-python==0.2.56
